Moved components to subdirectory
